+++ stores/mobx_triple/CHANGELOG.md
@@ -1,3 +1,7 @@
+ ## [1.1.0] - 2021-10-21
+
+- Update Triple
+
  ## [1.1.0] - 2021-03-04
 
 - Fix mobx version